Convert parameters of relation filters from float to double and apply static casts to float in the comparisons in the source. The reason for this is change is that it's not possible to set float parameters from Python steering files, not even when adding float to the variant list in trackFindingCDC/utilities/ParameterVariant. As soon as basf2 is executed when attempting to set the float parameters, they apparently are treated as double regardless, resulting in a segfault since only 4 byte for float are reserved in the class, but 8 byte are tried to be modified (at least this is my assumption). This commit fixes the issue. Note, though, that in general it's possible to set float parameters for regular basf2 modules that are NOT from TrackFindingCDC.

+++ tracking/ckf/pxd/filters/relations/include/AngularDistancePXDPairFilter.h
@@ -21,8 +21,8 @@ namespace Belle2 {
 
   private:
     /// Pre-filter relations in phi between seed states and hit states
-    float m_param_PhiRecoTrackToHitCut = 0.8;
+    double m_param_PhiRecoTrackToHitCut = 0.8;
     /// Pre-filter relations in phi between hit states
-    float m_param_PhiHitHitCut = 1.0;
+    double m_param_PhiHitHitCut = 1.0;
   };
 }
