+++ tracking/ckf/pxd/filters/relations/include/SensorPXDPairFilter.h
@@ -22,9 +22,9 @@ namespace Belle2 {
   private:
     /// Pre-filter relations in phi between seed states and hit states
     /// Here we compare RecoTrack -> mSoP -> getPos().Phi() with the phi value of the centre of the sensor the hit is on.
-    float m_param_PhiRecoTrackToHitCut = 0.8;
+    double m_param_PhiRecoTrackToHitCut = 0.8;
     /// Pre-filter relations in phi between hit states
     /// We compare hitA.sensorCenterPhi with hitB.sensorCenterPhi
-    float m_param_PhiHitHitCut = 1.0;
+    double m_param_PhiHitHitCut = 1.0;
   };
 }
